+++ packages/snjs/lib/Application/Dependencies/Dependencies.ts
@@ -168,6 +168,7 @@ import { Logger, isNotUndefined, isDeinitable, LoggerInterface } from '@standard
 import { EncryptionOperators } from '@standardnotes/encryption'
 import { AsymmetricMessagePayload, AsymmetricMessageSharedVaultInvite } from '@standardnotes/models'
 import { PureCryptoInterface } from '@standardnotes/sncrypto-common'
+import { AuthorizeVaultDeletion } from '@standardnotes/services/src/Domain/Vault/UseCase/AuthorizeVaultDeletion'
 
 export class Dependencies {
   private factory = new Map<symbol, () => unknown>()
@@ -249,6 +250,15 @@ export class Dependencies {
       return new ValidateVaultPassword(
         this.get<EncryptionService>(TYPES.EncryptionService),
         this.get<KeySystemKeyManager>(TYPES.KeySystemKeyManager),
+      )
+    })
+
+    this.factory.set(TYPES.AuthorizeVaultDeletion, () => {
+      return new AuthorizeVaultDeletion(
+        this.get<VaultLockService>(TYPES.VaultLockService),
+        this.get<ProtectionService>(TYPES.ProtectionService),
+        this.get<ChallengeService>(TYPES.ChallengeService),
+        this.get<ValidateVaultPassword>(TYPES.ValidateVaultPassword),
       )
     })
 
@@ -911,6 +921,7 @@ export class Dependencies {
         this.get<SendVaultDataChangedMessage>(TYPES.SendVaultDataChangedMessage),
         this.get<IsVaultOwner>(TYPES.IsVaultOwner),
         this.get<ValidateVaultPassword>(TYPES.ValidateVaultPassword),
+        this.get<AuthorizeVaultDeletion>(TYPES.AuthorizeVaultDeletion),
         this.get<InternalEventBus>(TYPES.InternalEventBus),
       )
     })