+++ packages/ui-services/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.29.12](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.29.11...@standardnotes/ui-services@1.29.12) (2023-08-09)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
+## [1.29.11](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.29.10...@standardnotes/ui-services@1.29.11) (2023-08-09)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
+## [1.29.10](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.29.9...@standardnotes/ui-services@1.29.10) (2023-08-09)
+
+**Note:** Version bump only for package @standardnotes/ui-services
+
 ## [1.29.9](https://github.com/standardnotes/app/compare/@standardnotes/ui-services@1.29.8...@standardnotes/ui-services@1.29.9) (2023-08-09)
 
 **Note:** Version bump only for package @standardnotes/ui-services